+++ admin/layouts/transaction_info.php
@@ -22,7 +22,7 @@ defined('_JEXEC') or die;
         <?php } else { ?>
             <tr class="cf-response-type">
                 <th colspan="2">
-                    <?php echo JHtml::_("crowdfundingbackend.trackId", $key); ?>
+                    <?php echo JHtml::_('crowdfundingbackend.trackId', $key); ?>
                 </th>
             </tr>
 
@@ -30,12 +30,16 @@ defined('_JEXEC') or die;
             <tr>
                 <th><?php echo $this->escape($k); ?></th>
                 <?php if (!is_array($v)) {?>
-                <td><?php echo (is_bool($v)) ? (!$v) ? "false" : "true" : $this->escape($v); ?></td>
+                <td><?php
+                    if (is_bool($v)) {
+                        echo (!$v) ? 'false' : 'true';
+                    }  else {
+                        echo $this->escape($v);
+                    }
+                    ?></td>
                 <?php } else { ?>
                 <td>
-                    <pre>
-                    <?php echo $this->escape(var_export($v, true)) ?>
-                    </pre>
+                    <pre><?php echo $this->escape(var_export($v, true)) ?></pre>
                 </td>
                 <?php } ?>