+++ libraries/lib_crowdfunding/Amount.php
@@ -261,7 +261,7 @@ class Amount
             $result = $this->parseAmount($this->value);
         }
 
-        return (float)$result;
+        return $result;
     }
 
     /**
@@ -292,7 +292,7 @@ class Amount
             return (float)$value;
         }
 
-        return (float)$value;
+        return $value;
     }
 
     /**
@@ -333,7 +333,6 @@ class Amount
         $format = explode('/', $format);
 
         if ((false !== $format) and count($format) > 0) {
-
             $value = (false !== strpos($value, ',')) ? $this->parse() : $value;
 
             $count = count($format);